requirements.txt
asgiref==3.3.4
class-registry==2.1.2
Django==3.2.4
django-filter==2.4.0
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
filters==1.3.2
Pillow==8.2.0
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
